faster power implementation

+++ 0x08-palindrome_integer/0-is_palindrome.c
@@ -6,7 +6,7 @@
  * @exp: exponant
  * Return: Result
  **/
-unsigned long ipow(int base, int exp)
+unsigned long ipow(unsigned long base, unsigned long exp)
 {
 	unsigned long result = 1;
